Adding update.version method
As I add features, the synced data vernacular might be changed/updated
to allow for new features. I need a means of verifying that sync setup
is as expected, so I'm adding an update file which will call various
updates until the most-current version setup is achieved.

Consider it a poor imitation of Django migrations.